EKF: apply IMU corrections at correct time horizon
Because the downsampling rotates delta velocities to the frame at the leading edge of the gyro data, the INS algorithm should rotate into earth frame after application of the delta angle.

NOTE: different group delays in the gyro and accelerometer sensors will cause inconsistencies in rotation that we cannot compensate for.