EKF: workaround for loss of accuracy in covariance prediction due to high spin rates
During high yaw rates the delta angle terms in the covariance prediction were causing the covariance matrix to not reflect the angular alignment between earth frame and body frame resulting in loss of roll and pitch reference during sustained high yaw rates. It is likely that this is caused by linearisation errors possibly in combination with rounding errors.

The short term solution is to zero the delta angle inputs.