estimator_interface: ensure that _time_last_imu is the latest timestamp
There are checks in the code such as this:
 _time_last_imu - _time_last_gps > (uint64_t)10e6
If the gps time is newer (by a couple of microseconds), the above check
will evaluate to true due to unsigned wrap-around logic.

Ensuring at the interface level that no other timestamp > _time_last_imu
is fed into the ecl fixes this. However this could potentially hide
timestamp problems.
For this reason, and since there was already a fix in
https://github.com/PX4/ecl/pull/450, the check is not added to the
external vision data.