EKF: Maintain flow sensor velocity limiting until flow data is no longer fused
This change fixes a vulnerability where if GPs use was gained, the velocity limiting was removed and the operator could fly the vehicle fast enough to cause the flow sensor to saturate and cause large velocity errors leading to rejection of GPS and loss of navigation.
velocity limiting is now matinee at all times when flow data is being fused, however the height limiting is removed when the vehicle is not completely reliant on flow data. This enables the vehicle to be flown up to an altitude where flow use ceases once GPS fusion has commenced.